# Apply binary optimisations