--- a/rebar.config
+++ b/rebar.config
@@ -29,6 +29,7 @@
 
 {profiles, [
     {test, [
+        {deps, [{proper, "1.4.0"}]},
         {erl_opts, [nowarn_export_all, nowarn_missing_spec, nowarn_missing_doc]},
         {eunit_opts, [verbose]},
         {covertool, [{coverdata_files, ["eunit.coverdata", "ct.coverdata"]}]},
@@ -42,7 +43,6 @@
     {warnings, [
         no_return,
         unmatched_returns,
-        error_handling,
         underspecs,
         unknown
     ]}