This is a port of the elixir implementation you can find at https://hexdocs.pm/elixir/Base.html

Encoding is 8-12x faster, decoding 2-4x.

Some benchmarks using benchee:

Hardware:
Operating System: macOS
CPU Information: Apple M3 Pro
Number of Available Cores: 12
Available memory: 36 GB
Elixir 1.18.2
Erlang 27.3
JIT enabled: true

Encode:
```
Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 5 s
reduction time: 5 s
parallel: 1
inputs: 17 bytes, 64 bytes
Estimated total run time: 1 min 8 s

Benchmarking candidate with input 17 bytes ...
Benchmarking candidate with input 64 bytes ...
Benchmarking original with input 17 bytes ...
Benchmarking original with input 64 bytes ...
Calculating statistics...
Formatting results...

----- With input 17 bytes -----
Name                ips        average  deviation         median         99th %
candidate       10.70 M       93.44 ns    ±62.56%          83 ns         125 ns
original         1.35 M      742.09 ns  ±2436.40%         625 ns        1000 ns

Comparison:
candidate       10.70 M
original         1.35 M - 7.94x slower +648.65 ns

Extended statistics:

Name              minimum        maximum    sample size                     mode
candidate            0 ns       96416 ns        19.06 M                    83 ns
original           333 ns    19270541 ns         5.67 M                   625 ns

Memory usage statistics:
Name         Memory usage
candidate           104 B
original            600 B - 5.77x memory usage +496 B

Reduction count statistics:
Name      Reduction count
candidate               8
original              138 - 17.25x reduction count +130

----- With input 64 bytes -----
Name                ips        average  deviation         median         99th %
candidate        6.35 M       0.157 μs    ±41.46%       0.166 μs        0.21 μs
original         0.56 M        1.78 μs   ±875.52%        1.63 μs        2.42 μs

Comparison:
candidate        6.35 M
original         0.56 M - 11.28x slower +1.62 μs

Extended statistics:
Name              minimum        maximum    sample size                     mode
candidate       0.0410 μs       59.54 μs        14.95 M                 0.167 μs
original             1 μs    11649.04 μs         2.63 M                  1.63 μs

Memory usage statistics:
Name         Memory usage
candidate           104 B
original            600 B - 5.77x memory usage +496 B

Reduction count statistics:
Name      Reduction count
candidate              17
original              435 - 25.59x reduction count +418
```

Decode:
```
Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 5 s
reduction time: 5 s
parallel: 1
inputs: 17 bytes before encoding, 64 bytes before encoding
Estimated total run time: 1 min 8 s

Benchmarking candidate with input 17 bytes before encoding ...
Benchmarking candidate with input 64 bytes before encoding ...
Benchmarking original with input 17 bytes before encoding ...
Benchmarking original with input 64 bytes before encoding ...
Calculating statistics...
Formatting results...

----- With input 17 bytes before encoding -----
Name                ips        average  deviation         median         99th %
candidate        5.58 M      179.18 ns  ±2111.19%         167 ns         291 ns
original         1.49 M      669.28 ns   ±255.79%         500 ns       12250 ns

Comparison:
candidate        5.58 M
original         1.49 M - 3.74x slower +490.10 ns

Extended statistics:

Name              minimum        maximum    sample size                     mode
candidate           83 ns     7605833 ns        13.12 M                   167 ns
original           250 ns      551292 ns         5.83 M                   500 ns

Memory usage statistics:
Name         Memory usage
candidate           184 B
original            104 B - 0.57x memory usage -80 B

Reduction count statistics:
Name      Reduction count
candidate               8
original               89 - 11.13x reduction count +81

----- With input 64 bytes before encoding -----
Name                ips        average  deviation         median         99th %
candidate        2.15 M        0.46 μs   ±197.53%        0.42 μs        0.96 μs
original         0.56 M        1.80 μs    ±25.90%        1.79 μs        2.29 μs

Comparison:
candidate        2.15 M
original         0.56 M - 3.87x slower +1.33 μs

Extended statistics:

Name              minimum        maximum    sample size                     mode
candidate         0.25 μs     1535.42 μs         7.30 M                  0.42 μs
original          1.04 μs      416.96 μs         2.53 M                  1.75 μs

Memory usage statistics:
Name         Memory usage
candidate           184 B
original            104 B - 0.57x memory usage -80 B

Reduction count statistics:
Name      Reduction count
candidate              17
original              314 - 18.47x reduction count +297
```